Add machine specific boot module
