+++ src/kernel/src/machine/arm/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod boot;
 pub mod gicv2;
 pub mod mmio;
 pub mod uart;
